bench: remove some duplication
